+++ example/lib/main.dart
@@ -4,6 +4,8 @@ import 'package:examples/images_grid.dart';
 import 'package:examples/simple.dart';
 import 'package:flutter/material.dart';
 
+import 'error.dart';
+
 void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
@@ -46,7 +48,18 @@ class MyApp extends StatelessWidget {
                 title: const Text('Data Stream example'),
                 onTap: () => Navigator.push(
                   context,
-                  MaterialPageRoute(builder: (_) => const DataStreamExample()),
+                  MaterialPageRoute(
+                    builder: (_) => const DataStreamExample(),
+                  ),
+                ),
+              ),
+              ListTile(
+                title: const Text('Data Stream error example'),
+                onTap: () => Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (_) => const SliverStreamErrorExample(),
+                  ),
                 ),
               ),
             ],